% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_scripts.R
\name{add_indexes_to_sqlite_db}
\alias{add_indexes_to_sqlite_db}
\title{Add indexes to key fields in SQLite db}
\usage{
add_indexes_to_sqlite_db(main_con)
}
\arguments{
\item{main_con}{A database connection object.}
}
\value{
None. The function updates the SQL database by adding indexes to key fields.
}
\description{
This function adds indexes to key fields in the SQLite database.
}
