% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_scripts.R
\name{add_program_col_to_sql_table}
\alias{add_program_col_to_sql_table}
\title{Add Program Column to SQL Table}
\usage{
add_program_col_to_sql_table(table, main_con)
}
\arguments{
\item{table}{A character string specifying the table name.}

\item{main_con}{A database connection object.}
}
\value{
None. The function updates the SQL table with an added "Program" column.
}
\description{
This function adds a "Program" column to a SQL table.
}
