% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_scripts.R
\name{add_program_col_to_tibble}
\alias{add_program_col_to_tibble}
\title{Add Program Column to Tibble}
\usage{
add_program_col_to_tibble(tibble)
}
\arguments{
\item{tibble}{A tibble dataframe.}
}
\value{
The tibble with an added "Program" column.
}
\description{
This function adds a "Program" column to a tibble.
}
