% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_SQLlite_files.R
\name{backup_SQLite_db}
\alias{backup_SQLite_db}
\title{Create a Monthly Backup of SQLite Database Files}
\usage{
backup_SQLite_db()
}
\value{
None. The function is called for its side effects.
}
\description{
This function creates a monthly backup of the SQLite database files. It ensures that a directory for storing monthly backups exists, creates a subdirectory for the current month, and copies the SQLite files to this subdirectory.
}
\examples{
backup_SQLite_db()
}
