% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{clean_UTF}
\alias{clean_UTF}
\title{Clean UTF strings}
\usage{
clean_UTF(x)
}
\arguments{
\item{x}{A string to be cleaned}
}
\value{
A cleaned string
}
\description{
This internal function cleans UTF-8 strings
}
\keyword{internal}
