% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{collapse_ids}
\alias{collapse_ids}
\title{Collapse IDs for API URL Construction}
\usage{
collapse_ids(df, id_col, paste_collapse)
}
\arguments{
\item{df}{A dataframe containing the IDs.}

\item{id_col}{A string specifying the column name that contains the IDs.}

\item{paste_collapse}{A string to use as the delimiter for collapsing the IDs.}
}
\value{
A single string of concatenated IDs separated by the specified delimiter.
}
\description{
This internal function collapses IDs from a specified column in a dataframe into a single string, separated by a specified delimiter. It is used for constructing API URLs.
}
\examples{
df <- data.frame(id = c("NCT1234567", "NCT987654321", NA, "NCT1234567"))
collapse_ids(df, "id", ",")
}
