% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{download_trial_info_wrapper}
\alias{download_trial_info_wrapper}
\title{Wrapper Function for All Updates}
\usage{
download_trial_info_wrapper(
  main_con = DBI::dbConnect(RSQLite::SQLite(),
    "inst/extdata/RSQLite_data/TrialTracker-db.sqlite"),
  trial_id_df = DBI::dbReadTable(main_con, "Trial_IDs"),
  dev_flag
)
}
\arguments{
\item{main_con}{A database connection object. Defaults to a connection to the TrialTracker SQLite database.}

\item{trial_id_df}{A dataframe containing trial IDs. Defaults to reading from the Trial_IDs table in the database.}

\item{dev_flag}{A development flag (logical TRUE / FALSE) toggling whether emails go to dev team only (TRUE) or dashboard users and dev team (FALSE).}
}
\value{
None. The function updates the database for all registries, updates PubMed tables, and sends email alerts.
}
\description{
This wrapper function updates the database for all registries and also updates PubMed tables and sends email alerts.
}
