% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{download_trial_info_wrapper_no_pm_or_email}
\alias{download_trial_info_wrapper_no_pm_or_email}
\title{Wrapper Function for All Updates (No PubMed or Email)}
\usage{
download_trial_info_wrapper_no_pm_or_email(
  main_con = DBI::dbConnect(RSQLite::SQLite(),
    "data/RSQLite_data/TrialTracker-db.sqlite"),
  trial_id_df = DBI::dbReadTable(main_con, "Trial_IDs")
)
}
\arguments{
\item{main_con}{A database connection object. Defaults to a connection to the TrialTracker SQLite database.}

\item{trial_id_df}{A dataframe containing trial IDs. Defaults to reading from the Trial_IDs table in the database.}

\item{db_connect_func}{A function for establishing a database connection (default: \code{DBI::dbConnect}).}

\item{db_read_func}{A function for reading a table from the database (default: \code{DBI::dbReadTable}).}
}
\value{
None. The function updates the database for all registries.
}
\description{
This wrapper function updates the database for all registries without updating PubMed tables or sending emails.
}
