% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_all_pubmed_dfs_for_all_programs_and_registries}
\alias{generate_all_pubmed_dfs_for_all_programs_and_registries}
\title{Generate All PubMed Dataframes for All Programs and Registries}
\usage{
generate_all_pubmed_dfs_for_all_programs_and_registries(
  registry_tables,
  prog_regexes,
  programs,
  main_con
)
}
\arguments{
\item{registry_tables}{A character vector of registry table names.}

\item{prog_regexes}{A character vector of regex patterns for programs.}

\item{programs}{A character vector of program names.}

\item{main_con}{A database connection object.}
}
\value{
None. The function generates PubMed dataframes and writes them to disk.
}
\description{
This function generates PubMed dataframes for all programs and registries.
}
