% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_all_pubmed_dfs_for_all_programs_one_registry}
\alias{generate_all_pubmed_dfs_for_all_programs_one_registry}
\title{Generate All PubMed Dataframes for All Programs in One Registry}
\usage{
generate_all_pubmed_dfs_for_all_programs_one_registry(
  registry,
  prog_regexes,
  programs,
  main_con
)
}
\arguments{
\item{registry}{A character string specifying the registry name.}

\item{prog_regexes}{A character vector of regex patterns for programs.}

\item{programs}{A character vector of program names.}

\item{main_con}{A database connection object.}
}
\value{
None. The function generates PubMed dataframes and writes them to disk.
}
\description{
This function generates PubMed dataframes for all programs in a specified registry.
}
