% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_change_files_for_nct_isrctn_eu}
\alias{generate_change_files_for_nct_isrctn_eu}
\title{Generate Change Files for NCT, ISRCTN, and EU Registries}
\usage{
generate_change_files_for_nct_isrctn_eu(
  prog_regexes,
  programs,
  registry_tables,
  main_con,
  daily_path
)
}
\arguments{
\item{prog_regexes}{A character vector of regex patterns for programs.}

\item{programs}{A character vector of program names.}

\item{registry_tables}{A character vector of registry table names.}

\item{main_con}{A database connection object.}

\item{daily_path}{A character string specifying the path to save the files.}
}
\value{
None. The function generates change files and writes them to disk.
}
\description{
This function generates change files for NCT, ISRCTN, and EU registries.
}
