% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_change_files_for_nihr}
\alias{generate_change_files_for_nihr}
\title{Generate Change Files for NIHR}
\usage{
generate_change_files_for_nihr(prog_regexes, programs, main_con, daily_path)
}
\arguments{
\item{prog_regexes}{A character vector of regex patterns for programs.}

\item{programs}{A character vector of program names.}

\item{main_con}{A database connection object.}

\item{daily_path}{A character string specifying the path to save the files.}
}
\value{
None. The function generates change files and writes them to disk.
}
\description{
This function loops through programs and generates change files for the NIHR registry.
}
