% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_API_functions.R
\name{generate_efetch_url}
\alias{generate_efetch_url}
\title{Generate EFetch URL}
\usage{
generate_efetch_url(pmid, api_object)
}
\arguments{
\item{pmid}{A character string specifying the PubMed ID.}

\item{api_object}{A character string specifying the API key.}
}
\value{
A character string containing the EFetch URL.
}
\description{
This function generates the EFetch URL for PubMed search.
}
