% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_email_alerts}
\alias{generate_email_alerts}
\title{Generate Email Alerts}
\usage{
generate_email_alerts(dev_flag)
}
\arguments{
\item{dev_flag}{A logical flag indicating if the email is for development purposes.}
}
\value{
None. The function generates and sends email alerts.
}
\description{
This function generates email alerts with attachments for trial tracking changes.
}
