% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_API_functions.R
\name{generate_esearch_url}
\alias{generate_esearch_url}
\title{Generate ESearch URL}
\usage{
generate_esearch_url(
  search_term,
  api_object,
  mindate = Sys.Date() - 1,
  maxdate = Sys.Date() - 1
)
}
\arguments{
\item{search_term}{A character string specifying the search term.}

\item{api_object}{A character string specifying the API key.}

\item{mindate}{The start date for the search. Defaults to yesterday's date.}

\item{maxdate}{The end date for the search. Defaults to yesterday's date.}
}
\value{
A character string containing the ESearch URL.
}
\description{
This function generates the ESearch URL for PubMed search.
}
