% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{generate_ISRCTN_df}
\alias{generate_ISRCTN_df}
\title{Generate ISRCTN Dataframe}
\usage{
generate_ISRCTN_df(ISRCTN_URL)
}
\arguments{
\item{ISRCTN_URL}{A character string specifying the URL to fetch the ISRCTN XML data.}
}
\value{
A tibble dataframe containing the ISRCTN trial data.
}
\description{
This function retrieves ISRCTN trial data from a given URL and returns it as a dataframe.
}
\examples{
\dontrun{
ISRCTN_URL <- "https://www.isrctn.com/api/query/format/who?q="
ISRCTN_ids <- c("ISRCTN27106947", "ISRCTN16033549", "ISRCTN17573805", "ISRCTN73327972")
ISRCTN_query <- paste0(paste(ISRCTN_ids, collapse = "\%20OR\%20"), "&limit=15")
df <- generate_ISRCTN_df(c(ISRCTN_URL, ISRCTN_query)
}
}
