% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{generate_ISRCTN_URL}
\alias{generate_ISRCTN_URL}
\title{Generate ISRCTN API URL}
\usage{
generate_ISRCTN_URL(ISRCTN_Id_Vector)
}
\arguments{
\item{ISRCTN_Id_Vector}{A character vector of ISRCTN trial IDs.}
}
\value{
A character string containing the URL for the ISRCTN API call.
}
\description{
This function creates a URL for an ISRCTN API call using a vector of ISRCTN trial IDs.
}
\examples{
generate_ISRCTN_URL(c("ISRCTN10985036", "ISRCTN25778550"))
}
