% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{generate_NCT_URL}
\alias{generate_NCT_URL}
\title{Generate URL for API Call to ClinicalTrials.gov}
\usage{
generate_NCT_URL(NCT_ID_Vec)
}
\arguments{
\item{NCT_ID_Vec}{A character vector of ClinicalTrials.gov Trial IDs.}
}
\value{
A string containing the URL for the API call.
}
\description{
This function generates a URL for an API call to ClinicalTrials.gov from a vector of Trial IDs. The URL is constructed to retrieve specific fields for the given trial IDs in JSON format.
}
\examples{
\dontrun{
trial_ids <- c("NCT03146143", "NCT00408044", "NCT01377987")
url <- generate_NCT_URL(trial_ids)
print(url)
}
}
