% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_nihr_comparison_table}
\alias{generate_nihr_comparison_table}
\title{Generate NIHR Comparison Table}
\usage{
generate_nihr_comparison_table(program, program_regex, main_con, daily_path)
}
\arguments{
\item{program}{A character string specifying the program name.}

\item{program_regex}{A regex pattern to filter Guideline numbers.}

\item{main_con}{A database connection object.}

\item{daily_path}{A character string specifying the path to save the file.}
}
\value{
None. The function generates a comparison table and writes it to disk.
}
\description{
This function generates a comparison table for a given program in the NIHR registry.
}
