% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_API_functions.R
\name{generate_pm_tibble_from_search_term_series}
\alias{generate_pm_tibble_from_search_term_series}
\title{Generate PubMed Tibble from Search Term Series}
\usage{
generate_pm_tibble_from_search_term_series(
  search_term_list,
  api_object,
  mindate = Sys.Date() - 1,
  maxdate = Sys.Date() - 1
)
}
\arguments{
\item{search_term_list}{A list of character strings specifying the search terms.}

\item{api_object}{A character string specifying the API key.}

\item{mindate}{The start date for the search. Defaults to yesterday's date.}

\item{maxdate}{The end date for the search. Defaults to yesterday's date.}
}
\value{
A tibble dataframe containing the PubMed results.
}
\description{
This function generates a tibble dataframe from a series of PubMed search terms.
}
