% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_pubmed_df_for_email_attachment}
\alias{generate_pubmed_df_for_email_attachment}
\title{Generate PubMed Dataframe for Email Attachment}
\usage{
generate_pubmed_df_for_email_attachment(
  main_con,
  registry_table,
  start_date,
  end_date,
  regex_pattern = "."
)
}
\arguments{
\item{main_con}{A database connection object.}

\item{registry_table}{A character string specifying the registry table name.}

\item{start_date}{The start date for generating the dataframe.}

\item{end_date}{The end date for generating the dataframe.}

\item{regex_pattern}{A regex pattern to filter Guideline numbers. Defaults to ".".}
}
\value{
A dataframe of PubMed results.
}
\description{
This function generates a PubMed dataframe for email attachments, based on specified dates.
}
