% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{generate_pubmed_results_from_search_terms_and_update_db_one_registry}
\alias{generate_pubmed_results_from_search_terms_and_update_db_one_registry}
\title{Update PubMed Tables for One Registry}
\usage{
generate_pubmed_results_from_search_terms_and_update_db_one_registry(
  registry,
  trial_id_df,
  main_con = DBI::dbConnect(RSQLite::SQLite(),
    "inst/extdata/RSQLite_data/TrialTracker-db.sqlite"),
  mindate = Sys.Date() - 1,
  maxdate = Sys.Date() - 1
)
}
\arguments{
\item{registry}{A character string specifying the registry (e.g., "ISRCTN", "NIHR").}

\item{trial_id_df}{The dataframe containing trial IDs.}

\item{main_con}{A database connection object. Defaults to a connection to the TrialTracker SQLite database.}

\item{mindate}{The start date for the PubMed search. Defaults to yesterday's date.}

\item{maxdate}{The end date for the PubMed search. Defaults to yesterday's date.}
}
\value{
None. The function updates the database with the PubMed results.
}
\description{
This function generates PubMed results from trial ids and updates the database for a specified registry.
}
