% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{generate_tt_email}
\alias{generate_tt_email}
\title{Generate Trial Tracker Email}
\usage{
generate_tt_email(
  program,
  attachments,
  dev_flag,
  users = readLines("secrets/users.csv"),
  devs = readLines("secrets/devs.csv")
)
}
\arguments{
\item{program}{A character string specifying the program name.}

\item{attachments}{A character vector of file paths to attach to the email.}

\item{dev_flag}{A logical flag indicating if the email is for the development version (dev team only).}

\item{users}{A character vector of user emails}

\item{devs}{A character vector of developer emails}
}
\value{
An email object.
}
\description{
This function generates an email to go out to users. Any changes found are detailed in attachments to the email.
}
