% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_API_functions.R
\name{get_efetch_results}
\alias{get_efetch_results}
\title{Get EFetch Results}
\usage{
get_efetch_results(pmid, api_object)
}
\arguments{
\item{pmid}{A character string specifying the PubMed ID.}

\item{api_object}{A character string specifying the API key.}
}
\value{
A list containing the EFetch results.
}
\description{
This function retrieves EFetch results from PubMed.
}
