% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_API_functions.R
\name{get_efetch_results_in_tibble_form}
\alias{get_efetch_results_in_tibble_form}
\title{Get EFetch Results in Tibble Form}
\usage{
get_efetch_results_in_tibble_form(pmid, api_object, search_term)
}
\arguments{
\item{pmid}{A character string specifying the PubMed ID.}

\item{api_object}{A character string specifying the API key.}

\item{search_term}{A character string specifying the search term.}
}
\value{
A tibble dataframe containing the PubMed results.
}
\description{
This function retrieves EFetch results and returns them in tibble form.
}
