% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_API_functions.R
\name{get_esearch_results}
\alias{get_esearch_results}
\title{Get ESearch Results}
\usage{
get_esearch_results(
  search_term,
  api_object,
  mindate = Sys.Date() - 1,
  maxdate = Sys.Date() - 1
)
}
\arguments{
\item{search_term}{A character string specifying the search term.}

\item{api_object}{A character string specifying the API key.}

\item{mindate}{The start date for the search. Defaults to yesterday's date.}

\item{maxdate}{The end date for the search. Defaults to yesterday's date.}
}
\value{
A list containing the ESearch results.
}
\description{
This function retrieves ESearch results from PubMed.
}
