% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{get_last_registry_entry_before_today}
\alias{get_last_registry_entry_before_today}
\title{Get Last Registry Entry Before Today}
\usage{
get_last_registry_entry_before_today(registry, main_con)
}
\arguments{
\item{registry}{A character string specifying the registry name.}

\item{main_con}{A database connection object.}
}
\value{
The last entry date before today as a Date object.
}
\description{
This function retrieves the last entry date before today for a specified registry.
}
