% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{load_attachments_and_send_email_alert_for_one_program}
\alias{load_attachments_and_send_email_alert_for_one_program}
\title{Load Attachments and Send Email Alert for One Program}
\usage{
load_attachments_and_send_email_alert_for_one_program(
  program,
  daily_path,
  dev_flag,
  smtp = smtp
)
}
\arguments{
\item{program}{A character string specifying the program name.}

\item{daily_path}{A character string specifying the path to the attachment files.}

\item{dev_flag}{A logical flag indicating if the email is for development purposes.}

\item{smtp}{An SMTP server object.}
}
\value{
None. The function sends an email alert with attachments.
}
\description{
This function attaches change files and sends an email alert for a specified program.
}
