% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{pull_change}
\alias{pull_change}
\title{Pull Change}
\usage{
pull_change(
  registry_table,
  main_con,
  start_date,
  end_date,
  group_cols = NULL,
  exclude_cols = NULL,
  regex_pattern = "."
)
}
\arguments{
\item{registry_table}{A character string specifying the registry table name.}

\item{main_con}{A database connection object.}

\item{start_date}{The start date for pulling changes.}

\item{end_date}{The end date for pulling changes.}

\item{group_cols}{A character vector of columns to group by. Defaults to NULL.}

\item{exclude_cols}{A character vector of columns to exclude. Defaults to NULL.}

\item{regex_pattern}{A regex pattern to filter Guideline numbers. Defaults to ".".}
}
\value{
A comparison dataframe of changes.
}
\description{
This function pulls changes from the registry table between specified dates.
}
