% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_helpers.R
\name{registry_key_from_col}
\alias{registry_key_from_col}
\title{Map Trial_Ids column name to registry key}
\usage{
registry_key_from_col(registry_col)
}
\arguments{
\item{registry_col}{One of "NCT_Ids","EU_Ids","ISRCTN_Ids","NIHR_Ids".}
}
\value{
Registry key: "NCT","EU","ISRCTN","NIHR".
}
\description{
Map Trial_Ids column name to registry key
}
\keyword{internal}
