% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_scripts.R
\name{remove_dups_and_overwrite_table}
\alias{remove_dups_and_overwrite_table}
\title{Remove Duplicates and Overwrite Table}
\usage{
remove_dups_and_overwrite_table(table_name, main_con)
}
\arguments{
\item{table_name}{A character string specifying the table name.}

\item{main_con}{A database connection object.}
}
\value{
None. The function updates the SQL table by removing duplicates.
}
\description{
This function removes duplicates from a table and writes back the table, overwriting the original.
}
