% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_scripts.R
\name{rename_ng_to_other}
\alias{rename_ng_to_other}
\title{Rename Program "NG" to "Other"}
\usage{
rename_ng_to_other(table_name, main_con)
}
\arguments{
\item{table_name}{A character string specifying the table name.}

\item{main_con}{A database connection object.}
}
\value{
None. The function updates the SQL table by renaming "NG" to "Other".
}
\description{
This function renames all instances of Program "NG" to "Other" in a specified table.
}
