% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_API_functions.R
\name{return_pm_df}
\alias{return_pm_df}
\title{Return PubMed Dataframe}
\usage{
return_pm_df(single_id_query_results, search_term)
}
\arguments{
\item{single_id_query_results}{A list containing the query results for a single ID.}

\item{search_term}{A character string specifying the search term.}
}
\value{
A tibble dataframe containing the PubMed results.
}
\description{
This function returns a dataframe from PubMed query results.
}
