% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{update_all_pubmed_tables}
\alias{update_all_pubmed_tables}
\title{Update All PubMed Tables}
\usage{
update_all_pubmed_tables(
  registries = c("NCT", "ISRCTN", "NIHR", "EU"),
  trial_id_df,
  main_con
)
}
\arguments{
\item{registries}{A character vector of registry names. Defaults to c("NCT", "ISRCTN", "NIHR", "EU").}

\item{trial_id_df}{A dataframe containing trial IDs.}

\item{main_con}{A database connection object.}
}
\value{
None. The function updates the PubMed tables in the database.
}
\description{
This wrapper function updates PubMed tables for all specified registries.
}
