% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{update_db}
\alias{update_db}
\title{Updates the SQLite Database with API Call Information}
\usage{
update_db(main_con, registry, DF)
}
\arguments{
\item{main_con}{A DBI connection object to the SQLite database.}

\item{registry}{A string specifying the name of the registry (table) to update.}

\item{DF}{A dataframe containing the data to be inserted into the database.}
}
\value{
None. This function is called for its side effects.
}
\description{
This function updates the specified SQLite database table with new data from an API call.
If the table does not exist, it creates the table and appends the data. If the table exists,
it appends the data only if there are no entries for the current date. If entries for the
current date exist, it overwrites the table with entries from previous dates and then appends
the new data.
}
\examples{
\dontrun{
main_con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DF <- data.frame(Query_Date = Sys.Date(), Value = 1:5)
update_db(main_con, "example_registry", DF)
}
}
