% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{write_changes_to_disk}
\alias{write_changes_to_disk}
\title{Write Changes to Disk}
\usage{
write_changes_to_disk(Change_DF, daily_path, DF_Name, prog_name = "")
}
\arguments{
\item{Change_DF}{A comparison dataframe of changes.}

\item{daily_path}{A character string specifying the path to save the file.}

\item{DF_Name}{A character string specifying the name of the dataframe.}

\item{prog_name}{A character string specifying the program name. Defaults to "".}
}
\value{
None. The function writes the changes to disk.
}
\description{
This function writes changes to disk as an Excel file.
}
