% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance_scripts.R
\name{write_prog_col_to_df}
\alias{write_prog_col_to_df}
\title{Write Program Column to DataFrame}
\usage{
write_prog_col_to_df(sql_table, main_con)
}
\arguments{
\item{sql_table}{A character string specifying the SQL table name.}

\item{main_con}{A database connection object.}
}
\value{
None. The function updates the SQL table with the "Program" column.
}
\description{
This function writes the "Program" column to a dataframe based on the "Guideline.number".
}
