% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert_functions.R
\name{write_pubmed_dfs_to_disk}
\alias{write_pubmed_dfs_to_disk}
\title{Write PubMed Dataframes to Disk}
\usage{
write_pubmed_dfs_to_disk(PM_DF, daily_path, PM_DF_Name)
}
\arguments{
\item{PM_DF}{A dataframe of PubMed results.}

\item{daily_path}{A character string specifying the path to save the file.}

\item{PM_DF_Name}{A character string specifying the name of the PubMed dataframe.}
}
\value{
None. The function writes the PubMed dataframes to disk.
}
\description{
This function writes PubMed dataframes to disk as CSV files.
}
