% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trial_info_functions.R
\name{generate_NCT_DF}
\alias{generate_NCT_DF}
\title{Generate Formatted Tibble from ClinicalTrials.gov API Call URL}
\usage{
generate_NCT_DF(url)
}
\arguments{
\item{url}{A string containing the URL for the API call to ClinicalTrials.gov.}
}
\value{
A tibble containing the formatted data from the API response.
}
\description{
This function returns a formatted tibble from a given ClinicalTrials.gov API call URL. The tibble includes various fields such as NCT ID, study ID, condition, brief title, acronym, overall status, and more.
}
\examples{
\dontrun{
api_url <- paste0(
  "https://clinicaltrials.gov/api/v2/studies?format=json&filter.ids=",
  "NCT03146143\%7CNCT00408044"
)

api_fields <- paste0(
  "&fields=",
  "NCTId\%7COrgStudyId\%7CCondition\%7CBriefTitle\%7CAcronym\%7COverallStatus\%7C",
  "PrimaryCompletionDate\%7CCompletionDate\%7CResultsFirstSubmitDate\%7C",
  "ResultsFirstPostDate\%7CLastUpdatePostDate\%7CSeeAlsoLinkURL",
  "&pageSize=1000"
)

df <- generate_NCT_DF(c(api_url, api_fields))
print(df)
}
}
